+++ pkg/meta/model.go
@@ -9,7 +9,7 @@ type ObjectMeta struct {
 	//
 	// Populated by the system.
 	// Read-only.
-	ID uint64 `json:"id,omitempty" gorm:"primary_key;AUTO_INCREMENT;column:id"`
+	ID int `json:"id,omitempty" gorm:"primary_key;AUTO_INCREMENT;column:id"`
 
 	// CreatedAt is a timestamp representing the server time when this object was
 	// created. It is not guaranteed to be set in happens-before order across separate operations.
